requirements.txt
numpy==1.19.2
tensorflow==2.1.0
matplotlib==3.3.4
pyqt==5.9.2
